Cancel timer for `Ember.run.later` in initializer.
This isn't the optimal fix: that's still to come, in the form of moving
this out of an initializer and into `app/router.js` instead. However,
this is a backwards-compatible stopgap which resolves a number
of test failures which can arise if there are any bad interactions
with other `Router` reopening or extension.